+++ templates/react-native/screen/Screen.test.tsx
@@ -3,41 +3,46 @@ import * as React from 'react';
 import Screen from '../Screen';
 
 import renderer from 'react-test-renderer';
-import { render, fireEvent } from 'react-native-testing-library';
-
-const createTestProps = (props: Object) => ({
-  navigation: {
-    navigate: jest.fn(),
-    goBack: jest.fn(),
-  },
-  ...props,
-});
+import { render, fireEvent, RenderAPI } from 'react-native-testing-library';
 
 let props: any;
-let testing: any;
-let component: any;
+let component: React.ReactElement;
+let testingLib: RenderAPI;
 
-describe('[Screen]', () => {
-  beforeAll(() => {
-    props = createTestProps({});
+const createTestProps = (obj: object) => ({
+  navigation: {
+    navigate: jest.fn(),
+  },
+  ...obj,
+});
+
+describe('[Screen] screen', () => {
+  beforeEach(() => {
+    props = createTestProps({ });
+    component = (
+      <Screen {...props} />
+    );
   });
 
   it('renders without crashing', () => {
-    const rendered: renderer.ReactTestRendererJSON = renderer.create(<Screen />).toJSON();
+    const rendered = renderer.create(component).toJSON();
     expect(rendered).toMatchSnapshot();
     expect(rendered).toBeTruthy();
   });
-});
-
-describe('[Screen] Interaction', () => {
-  beforeEach(() => {
-    props = createTestProps({});
-    component = <Screen {...props} />;
-    testing = render(component);
-  });
 
   it('should render [Text] with value "myText"', () => {
-    const textInstance: renderer.ReactTestInstance = testing.getByTestId('myText');
+    const textInstance: renderer.ReactTestInstance = testingLib.getByTestId('myText');
     expect(textInstance.props.children).toEqual('dooboolab');
+  });
+
+  describe('interactions', () => {
+    beforeEach(() => {
+      testingLib = render(component);
+    });
+
+    it('should simulate onClick', () => {
+      // fireEvent(testingLib.getByTestId('btn'), 'click');
+      // expect(cnt).toBe(2);
+    });
   });
 });