+++ templates/react-native/shared/Shared.tsx
@@ -4,7 +4,7 @@ import {
   Text,
 } from 'react-native';
 
-import styled from 'styled-components';
+import styled from 'styled-components/native';
 
 const Container = styled.View`
   flex: 1;